src/django_backend/abapong/requirements.txt
Django==2.1.1
Pillow==4.1.0
djangorestframework==3.6.4
uWSGI==2.0.17
psycopg2==2.6
docutils==0.12
Markdown==2.6.7
django-versatileimagefield==1.7.1
requests==2.13.0
django-rest-framework-roles==0.5
djangorestframework-filters==0.10.2
python-dateutil==2.6.1
requests==2.21.0